fix #379 : MotionManager 파일로 분리
