+++ Outline/Outline/Source/Utils/Component/Card/BigCard.swift
@@ -5,7 +5,6 @@
 //  Created by hyunjun on 11/18/23.
 //
 
-import CoreMotion
 import SwiftUI
 import MapKit
 
@@ -187,32 +186,6 @@ struct BigCard<Content: View>: View {
         let absRotationAngle = abs(rotationAngle)
         let isOddRotation = Int((absRotationAngle + 90) / 180) % 2 != 0
         isFrontside = (isOddRotation && isFliped) || (!isOddRotation && !isFliped)
-    }
-}
-
-final class MotionManager: ObservableObject {
-    @Published var pitch: Double = 0.0
-    @Published var roll: Double = 0.0
-    
-    private var manager: CMMotionManager
-    
-    init() {
-        self.manager = CMMotionManager()
-        self.manager.deviceMotionUpdateInterval = 1/60
-        self.manager.startDeviceMotionUpdates(to: .main) { (motionData, error) in
-            guard error == nil else {
-                print(error!)
-                return
-            }
-            
-            if let motionData = motionData {
-                withAnimation {
-                    self.pitch = abs(motionData.attitude.pitch)
-                    self.roll = abs(motionData.attitude.roll)
-                }
-            }
-        }
-        
     }
 }
